Move org.gradle.caching.configuration.internal to :core
