+++ subprojects/core/src/main/java/org/gradle/caching/configuration/internal/package-info.java
