+++ subprojects/core/src/test/groovy/org/gradle/caching/configuration/internal/DefaultBuildCacheConfigurationTest.groovy
